% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bulk_Read_Counts}
\alias{Bulk_Read_Counts}
\title{Parekh et al. 2016: Gene Expression Matrices}
\format{An object of class \code{data.frame} with 57773 rows and 10 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75823}.
}
\usage{
data(Bulk_Read_Counts)
}
\description{
250 ng of Universal Human Reference RNA (UHRR; Agilent Technologies; catalog #740000) and ERCC spike-in control mix I (Life Technologies) were used and cDNA was synthesized as described in the Smart-Seq2 protocol from Picelli et al. 2013.\cr
Base-calls were performed with Illumina bcl2fastq.\cr
All cDNA reads were aligned with NextGenMap 0.4.12. Smart-seq2 reads were mapped to GRCh37 (hg19) and ERCC synthetic spike-in reference.
Reads were assigned to genes using FeatureCount method in Rsubread package.\cr
}
\keyword{datasets}
