% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param.R
\name{estimateNBParam}
\alias{estimateNBParam}
\title{Estimate simulation parameters}
\usage{
estimateNBParam(countData, cData=NULL, design=NULL,
RNAseq=c("bulk", "singlecell"),
estFramework=c('edgeR', 'DESeq2', 'MatchMoments'),
sigma=1.96)
}
\arguments{
\item{countData}{is a count \code{matrix}. Rows correspond to genes, columns to samples.}

\item{cData}{A \code{data.frame} with at least a single column. Rows of \code{colData} \strong{must correspond} to columns of \code{countData}. Default is \code{NULL}, i.e. the dispersion estimation is 'blind' to sample information, see \code{\link[DESeq2]{varianceStabilizingTransformation}}.}

\item{design}{A \code{formula} which expresses how the counts for each gene depend on the variables in colData. Designs with multiple covariates and/or interactions are supported. Default is \code{NULL}, i.e. no design information is considered.}

\item{RNAseq}{is a character value: "bulk" or "singlecell".}

\item{estFramework}{is a character value: "edgeR", "DESeq2" and "MatchMoments".
"edgeR" or "DESeq2" employs the edgeR or DESeq2 style mean and dispersion estimation, respectively. For details, please consult \code{\link[edgeR]{estimateDisp}} and \code{\link[DESeq2]{estimateDispersions}}.
"MatchMoments" employs moments matching technique for of mean, dispersion and dropout estimation.}

\item{sigma}{The variability band width for mean-dispersion loess fit defining the prediction interval for read cound simulation. Default is 1.96, i.e. 95\% interval. For more information see \code{\link[msir]{loess.sd}}.}
}
\value{
List with the following vectors
\item{seqDepth}{Library size, i.e. total number of reads per library}
\item{means}{Mean normalized read counts per gene.}
\item{dispersion}{Dispersion estimate per gene.}
\item{common.dispersion}{The common dispersion estimate over all genes.}
\item{size}{Size parameter of the negative binomial distribution, i.e. 1/dispersion.}
\item{p0}{Probability that the count will be zero per gene.}
\item{meansizefit}{A loess fit relating log2 mean to log2 size for use in simulating new data (\code{\link[msir]{loess.sd}}).}
\item{meandispfit}{A fit relating log2 mean to log2 dispersion used for visualizing mean-variance dependency (\code{\link[msir]{loess.sd}}).}
\item{p0.cut}{The knee point of meanp0fit. Log2 mean values above that value have virtually no dropouts.}
\item{grand.dropout}{The percentage of empty entries in the count matrix.}
\item{sf}{The estimated library size factor per sample.}
\item{totalS,totalG}{Number of samples and genes provided.}
\item{estS,estG}{Number of samples and genes for which parameters can be estimated.}
\item{RNAseq}{The type of RNAseq: bulk or single cell.}
\item{estFramework}{The estimation framework for NB parameters.}
\item{sigma}{The width of the variability band.}
}
\description{
This function estimates and returns parameters needed for power simulations assuming a negative binomial read count distribution.
}
\examples{
\dontrun{
## simulating single cell RNA-seq experiment
ngenes <- 10000
ncells <- 100
true.means <- 2^runif(ngenes, 3, 6)
true.dispersions <- 3 + 100/true.means
sf.values <- 2^rnorm(ncells, sd=0.5)
sf.means <- outer(true.means, sf.values, '*')
cnts <- matrix(rnbinom(ngenes*ncells,
mu=sf.means, size=1/true.dispersions),
ncol=ncells)
## estimating negative binomial parameters
estparam <- estimateNBParam(cnts, RNAseq = 'singlecell',
estFramework = 'MatchMoments', sigma=1.96)
plotNBParam(estparam)

## simulating bulk RNA-seq experiment
ngenes <- 10000
nsamples <- 10
true.means <- 2^rnorm(ngenes, mean=8, sd=2)
true.dispersions <- rgamma(ngenes, 2, 6)
sf.values <- rnorm(nsamples, mean=1, sd=0.1)
sf.means <- outer(true.means, sf.values, '*')
cnts <- matrix(rnbinom(ngenes*nsamples,
mu=sf.means, size=1/true.dispersions),
ncol=nsamples)
## estimating negative binomial parameters
estparam <- estimateNBParam(cnts, RNAseq = 'bulk',
estFramework = 'MatchMoments', sigma=1.96)
plotNBParam(estparam)
}
}
\author{
Beate
}
