% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_lengths}
\alias{gene_lengths}
\title{Ziegenhain et al. 2017: Gene Lengths}
\format{An object of class \code{integer} of length 39016.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75790}.
}
\usage{
gene_lengths
}
\description{
ENSEMBL gene length annotation for SCRBseq and Smartseq2 library preparation.
}
\keyword{datasets}
