% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneLengths_hg19}
\alias{GeneLengths_hg19}
\title{Parekh et al. 2016: Gene Lengths}
\format{A named vector of gene lengths.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75823}.
}
\usage{
data(GeneLengths_hg19)
}
\description{
ENSEMBL gene length annotation for gene expression quantification in Parekh et al. 2016.
}
\keyword{datasets}
