% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneLengths_mm10}
\alias{GeneLengths_mm10}
\title{Ziegenhain et al. 2017: Gene Lengths}
\format{
A named vector of gene lengths.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75790}.
}
\usage{
data(GeneLengths_mm10)
}
\description{
ENSEMBL gene length annotation for gene expression quantification in Ziegenhain et al. 2017.
}
\keyword{datasets}
