% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kolodziejczk_cnts}
\alias{kolodziejczk_cnts}
\title{Kolodziejczk et al. 2015: Gene Counts}
\format{An object of class \code{matrix} with 10000 rows and 869 columns.}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-2600/}.
}
\usage{
kolodziejczk_cnts
}
\description{
869 mouse embryonic stem cells under three different conditions (cnts data object):
serum + LIF (242 cells), 2i + LIF (433 cells) and alternative 2i + LIF (194 cells).
Single cell RNA-Seq was performed using Fluidigm C1 system and
libraries were generated using Nextera XT (Illumina) kit.
The read count matrix was downsampled to 10000 genes.
}
\keyword{datasets}
