% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kolodziejczk_param}
\alias{kolodziejczk_param}
\title{Kolodziejczk et al. 2015: Parameter Estimation}
\format{An object of class \code{list} of length 26.}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-2600/}.
}
\usage{
kolodziejczk_param
}
\description{
869 mouse embryonic stem cells under three different conditions (cnts data object):
serum + LIF (242 cells), 2i + LIF (433 cells) and alternative 2i + LIF (194 cells).
Single cell RNA-Seq was performed using Fluidigm C1 system and
libraries were generated using Nextera XT (Illumina) kit.
We reduced the data set to standard serum + LIF cultured cells for
parameter estimation.
}
\keyword{datasets}
