% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kolodziejczk_simDE}
\alias{kolodziejczk_simDE}
\title{Kolodziejczk et al. 2015: Simulation Result}
\format{An object of class \code{list} of length 14.}
\source{
\url{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-2600/}.
}
\usage{
kolodziejczk_simDE
}
\description{
869 mouse embryonic stem cells under three different conditions (cnts data object):
serum + LIF (242 cells), 2i + LIF (433 cells) and alternative 2i + LIF (194 cells).
Single cell RNA-Seq was performed using Fluidigm C1 system and
libraries were generated using Nextera XT (Illumina) kit.
We reduced the data set to standard serum + LIF cultured cells for
parameter estimation. For the simulation 20% of 10000 simulated genes were differentially expressed with a broad symmetrical gamma distribution of log2 fold changes. For DE testing, limma trend was used.
}
\keyword{datasets}
