% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotCounts}
\alias{plotCounts}
\title{Visualize simulated counts}
\usage{
plotCounts(simCounts, Distance, Scale, DimReduce, verbose = T)
}
\arguments{
\item{simCounts}{The output of \code{\link{simulateCounts}}.}

\item{Distance}{The (dis-)similarity measure to be used. This can be "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski" for distance measures and "spearman", "pearson" or "kendall" for correlation measures converted into distances, respectively. For more information, see \code{\link[stats]{dist}} and \code{\link[stats]{cor}}.}

\item{Scale}{A logical vector indicating whether to use scaled log2 transformed counts or not.}

\item{DimReduce}{The dimension reduction approach to be used. This can be "MDS" \code{\link[stats]{cmdscale}}, "PCA" \code{\link[stats]{prcomp}}, "t-SNE" \code{\link[Rtsne]{Rtsne}}, "ICA" \code{\link[fastICA]{fastICA}} or "LDA" \code{\link[MASS]{lda}}.}

\item{verbose}{Logical value to indicate whether to print function information. Default is \code{TRUE}.}
}
\value{
A ggplot object.
}
\description{
This function performs multidimensional scaling of the simulated counts from the pairwise sample distances using variable genes (i.e. variance unequal to zero). Prior to distance calculation, the counts are normalized using the simulated size factors and log2 transformed. In the plot, the samples are annotated by phenotype and batch, if present.
}
\examples{
\dontrun{
## not yet
}
}
\author{
Beate Vieth
}
