% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotNBParam}
\alias{plotNBParam}
\title{Visualize}
\usage{
plotNBParam(estParam.out, annot=TRUE)
}
\arguments{
\item{estParam.out}{The output of \code{\link{estimateNBParam}} or a string specifying the name of precalculated estimates (see details).}

\item{annot}{A logical vector. If \code{TRUE}, a short figure legend is included.}
}
\value{
A ggplot object.
}
\description{
This function plots the results of the parameter estimation. This includes the absolute and relative sequencing depth (i.e. library size factor). Also the mean-dispersion relationship with loess fit for simulations is visualized. Lastly the mean-dropout rate is presented as a smooth scatter plot.
}
\details{
Precalculated negative binomial parameters for simulations are available on \url{}.
}
\examples{
\dontrun{
## for example see \\code{\\link{insilicoNBParam}}
}
}
\author{
Beate Vieth
}
