% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotSpike}
\alias{plotSpike}
\title{Visualize distributional characteristics of spike-ins}
\usage{
plotSpike(estSpike, annot=TRUE)
}
\arguments{
\item{estSpike}{The output of \code{\link{estimateParam}}.}

\item{annot}{A logical vector. If \code{TRUE}, a short figure legend is included.}
}
\value{
A ggplot object.
}
\description{
This function plots the results of the parameter estimation for spike-ins. This includes the absolute and relative sequencing depth (i.e. library size factor), a calibration curve as well as the capture efficiency given as a binomial regression.
}
\examples{
\dontrun{
#' ## batch annotation
data(scrbseq_spike_cnts)
data(scrbseq_spike_info)
batch_info <- data.frame(Batch = ifelse(grepl(pattern = "SCRBseqA_",
colnames(scrbseq_spike_cnts)), "A", "B"),
row.names = colnames(scrbseq_spike_cnts))
## spike information table
spike_info <- scrbseq_spike_info[-1,]
## estimation
spike_param <- estimateSpike(spikeData = scrbseq_spike_cnts,
spikeInfo = spike_info,
MeanFragLength = NULL,
batchData = batch_info,
normalisation = 'depth')
## plotting
plotSpike(estSpike = spike_param, annot=TRUE)
}
}
\author{
Beate Vieth
}
