% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrbseq_spike_cnts}
\alias{scrbseq_spike_cnts}
\title{Ziegenhain et al. 2017: SCRBseq Spike-Ins Counts}
\format{An object of class \code{data.frame} with 92 rows and 58 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75790}.
}
\usage{
scrbseq_spike_cnts
}
\description{
Spike-ins read counts for SCRBseq library preparation.
}
\keyword{datasets}
