% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scrbseq_spike_info}
\alias{scrbseq_spike_info}
\title{Ziegenhain et al. 2017: SCRBseq Spike-Ins Information}
\format{An object of class \code{data.frame} with 92 rows and 3 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75790}.
}
\usage{
scrbseq_spike_info
}
\description{
Spike-ins information table with molecules and spike-ins transcript lengths for SCRBseq library preparation.
}
\keyword{datasets}
