% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smartseq2_gene_cnts}
\alias{smartseq2_gene_cnts}
\title{Ziegenhain et al. 2017: Smartseq2 Gene Counts}
\format{An object of class \code{data.frame} with 10000 rows and 58 columns.}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75790}.
}
\usage{
smartseq2_gene_cnts
}
\description{
J1 mouse embryonic stem cells (cultured in medium) were facs sorted and processed in two batches.
Single cell RNA-Seq was performed using Smartseq2 protocol and libraries were generated using Nextera XT (Illumina) kit.
Furthermore, spike-ins (ERCC) were added.
The read count matrix was downsampled to 10000 genes.
}
\keyword{datasets}
