% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SpikeInfo}
\alias{SpikeInfo}
\title{Ziegenhain et al. 2017: Spike-Ins Information}
\format{
An object of class \code{"data.frame"}, with spike-in IDs, number of spike-in molecules and lengths of spike-ins.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75790}.
}
\usage{
data(CELseq2_SpikeInfo)
data(MARSseq_SpikeInfo)
data(SCRBseq_SpikeInfo)
data(SmartSeq_SpikeInfo)
data(SmartSeq2_SpikeInfo)
}
\description{
Spike-ins information table with molecules and spike-ins transcript lengths per library preparation protocol.
}
\keyword{datasets}
