% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{evaluateDist}
\alias{evaluateDist}
\title{Model Diagnostics for RNAseq Data}
\usage{
evaluateDist(countData, batchData =NULL,
spikeData = NULL, spikeInfo = NULL,
Lengths = NULL, MeanFragLengths = NULL,
RNAseq, Normalisation,
frac.genes=1, min.meancount = 0.1,
max.dropout=0.7, min.libsize=1000,
verbose = TRUE)
}
\arguments{
\item{countData}{is a count matrix (row=gene, column=sample).
Please provide the measurements of one group only, e.g. the control group.}

\item{batchData}{is a \code{data.frame} for batch annotation.
Rows correspond to samples. The first column should contain the batches, e.g. 'a', 'b', 'c', etc..
This is only used for the normalisation.}

\item{spikeData}{is a count \code{matrix}.
Rows correspond to spike-ins, columns to samples.
The order of columns should be the same as in the \code{countData}.}

\item{spikeInfo}{is a molecule count \code{matrix} of spike-ins.
Rows correspond to spike-ins. The order of rows should be the same as in the \code{spikeData}.
The column names should be 'SpikeID' and 'SpikeInput' for molecule counts of spike-ins.}

\item{Lengths}{is a numeric vector of transcript lengths with the same length and order as the rows in countData.
This variable is only used for internal TPM calculations if Census normalization is specified.}

\item{MeanFragLengths}{is a numeric vector of mean fragment lengths with the same length as columns in countData.
This variable is only used for internal TPM calculations if Census normalization is specified.}

\item{RNAseq}{is a character value: "bulk" or "singlecell".}

\item{Normalisation}{is a character value: 'TMM', 'MR', 'PosCounts', 'UQ', 'scran', 'Linnorm',
'SCnorm', 'RUV', 'Census', 'depth', 'none'.
For more information, please consult the details section of \code{\link{estimateParam}}.}

\item{frac.genes}{The fraction of genes to calculate goodness of fit statistics, default is 1, i.e. for all genes.}

\item{min.meancount}{The minimum mean normalised count per gene if fraction of genes are defined. Default is \code{0.1}.}

\item{max.dropout}{The maximal percentage of zero expression per gene. Genes with more than \code{max.dropout} will be remove. Default is \code{0.7}, i.e. genes with more than 70\% dropouts.}

\item{min.libsize}{The minimum raw read counts per sample, default is \code{1000}.}

\item{verbose}{Logical value to indicate whether to show progress report of simulations.}
}
\value{
List object with the results of goodness of fit and estimated parameters:
\item{edgeR}{Goodness-of-fit statistic, degrees of freedom and associated p-value using the deviance and residual degrees of freedom from \code{\link[edgeR]{glmFit}}. Furthermore, the AIC of the edgeR model fit using the residuals of \code{\link[edgeR]{zscoreNBinom}}.}
\item{GOF}{The fitting results per distribution, including loglikelihood, goodness-of-fit statistics, AIC and predicted number of zeroes. The following distributions were considered: Poisson, negative binomial, zero-inflated poisson and negative binomial following the 'standard' (i.e. \code{\link[stats]{glm}}, \code{\link[MASS]{glm.nb}} and \code{\link[pscl]{zeroinfl}} implementation) and fitdist approach (see \code{\link[fitdistrplus]{fitdist}}) and Beta-Poisson following Marioni or Hemberg parameterisation. Furthermore, model fit comparison by LRT for nested and Vuong Test for non-nested models.}
\item{Estimates}{The estimated parameters of distribution fitting.}
\item{ObservedZeros}{The number of zeroes and dropout rate per gene.}
}
\description{
With this function, the user can determine goodness of fit for each gene.
}
\examples{
\dontrun{
## using example data set
data(kolodziejczk_cnts)
evaldist <- evaluateDist(countData = kolodziejczk_cnts,
RNAseq = "singlecell", Normalisation="scran",
frac.genes=1, min.meancount = 0.1,
max.dropout=0.7, min.libsize=1000,
verbose = TRUE)
plotEvalDist(evaldist, annot = TRUE)
}
}
\author{
Beate Vieth
}
