% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalDist.R
\name{evaluateDist}
\alias{evaluateDist}
\title{Model Diagnostics for RNAseq Data Set}
\usage{
evaluateDist(cnts, RNAseq, ncores=1, nsims=1,
frac.genes=1, min.meancount=1, min.libsize=1000)
}
\arguments{
\item{cnts}{is a count matrix (row=gene, column=sample). Pprovide the measurements of one group only, e.g. the control group.}

\item{RNAseq}{Character vector for "singlecell" or "bulk".}

\item{ncores, }{number of cores for parallel computing, default is 1.}

\item{nsims}{Number of simulations for MC p-value calculations, default is 1.}

\item{frac.genes}{The fraction of genes to calculate goodness of fit statistics, default is 1, i.e. for all genes.}

\item{min.meancount}{The minimum raw mean count per gene, default is 1.}

\item{min.libsize}{The minimum raw read counts per sample, default is 1000.}
}
\value{
A list object with the results of goodness of fit and estimated parameters.
}
\description{
With this function, the user can determine goodness of fit for each gene.
}
\examples{
\dontrun{
## simulating read count matrix
ngenes <- 10000
ncells <- 100
## NB genes with dropouts:
ZNB.genes <- 2^rgamma(ngenes*0.45, 2, 2)
## NB genes with high expression:
NB.genes <- 2^runif(ngenes*0.45, 9, 12)
## Poisson genes:
P.genes <- 2^runif(ngenes*0.1, 3, 6)
## all means:
true.means <- c(ZNB.genes, NB.genes, P.genes)
sf.values <- rnorm(ncells, mean=1, sd=0.1)
sf.means <- outer(true.means, sf.values, '*')
true.dispersions <- 3 + 100/true.means[1:round(ngenes*0.9)]
## count matrix:
cnts.NB <- matrix(rnbinom(ngenes*0.9*ncells,
mu=sf.means[1:round(ngenes*0.9)],
size=1/true.dispersions),
ncol=ncells)
cnts.P <- matrix(rpois(ngenes*0.1*ncells,
lambda=sf.means[round(ngenes*0.9)+1:ngenes]),
ncol=ncells)
cnts <- rbind(cnts.NB, cnts.P)
## evaluate distribution fitting:
evaldistres <- evaluateDist(cnts=cnts, RNAseq='singlecell',
ncores=1, nsims=1, frac.genes=0.25,
min.meancount=0.1, min.libsize=1000)
## plot the evaluation results:
plotEvalDist(evalDist=evaldistres, annot=TRUE)
}
}
\author{
Beate Vieth
}
