% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{evaluateROC}
\alias{evaluateROC}
\title{Receiver operator characteristics from simulation results}
\usage{
evaluateROC(simRes,
alpha.type=c("adjusted","raw"),
MTC=c('BY', 'BH', 'Storey', 'IHW',
'holm', 'hochberg', 'hommel', 'bonferroni'),
target.by=c("lfc", "effectsize"),
delta=0)
}
\arguments{
\item{simRes}{The result from \code{\link{simulateDE}}.}

\item{alpha.type}{A string to represent the way to call DE genes.
Available options are \code{"adjusted"} i.e. applying multiple testing correction and
\code{"raw"} i.e. using p-values. Default is \code{"adjusted"}.}

\item{MTC}{Multiple testing correction method to use. Available options are
1) see \link[stats]{p.adjust.methods},
2) Storey's qvalue see \link[qvalue]{qvalue} and
3) Independent Hypothesis Weighting considering mean expression as covariate (see \link[IHW]{ihw}).
Default is \code{BY}, i.e. Benjamini-Yekutieli FDR correction method.}

\item{target.by}{A string to specify the method to define "biologically important" DE genes.
Available options are (1) \code{"lfc"}: interesting genes are defined by absolute log2 fold changes.
(2) \code{"effectsize"}: interesting genes are defined by
absolute log2 fold changes divided by the square root of 1/(mean+dispersion).}

\item{delta}{A threshold used for defining "biologically important" genes.
Genes with absolute log2 fold changes (when target.by is "lfc")
or effect sizes (when target.by is "effectsize") greater than this value
are deemed DE in error rates calculations.
If \code{delta=0} then no threshold is applied. See "Details" for more description.}
}
\value{
A list with the following entries:
\item{ROCR-Prediction}{The output of \code{\link[ROCR]{prediction}} split up by group comparisons into lists.}
\item{ROCR-Performance}{The output of \code{\link[ROCR]{performance}} calculating TPR and FPR for ROC curve.
This is also split up by group comparisons into lists.}
\item{ROCR-AUC}{The output of \code{\link[ROCR]{performance}} calculating AUC of ROC curve.
split up by group comparisons into lists.}
\item{n1,n2}{Sample sizes per group.
This is taken from the simulation options.}
\item{target.by}{The input method to define "biologically important" DE genes,
either by log fold change or effect size.}
\item{delta}{The input delta for biologically important genes.
If delta=0, all target.by will be considered.}
}
\description{
This function takes the simulation output from \code{\link{simulateDE}}
and computes receiver operator characteristics (ROC) and area under the curve values (AUC).
}
\examples{
\dontrun{
data(kolodziejczk_simDE)
eval.roc <- evaluateROC(simRes = kolodziejczk_simDE)
}
}
\seealso{
\code{\link{estimateParam}} for negative binomial parameters,
\code{\link{SimSetup}} and
\code{\link{DESetup}} for setting up simulation parameters and
\code{\link{simulateDE}} for simulating differential expression.
}
\author{
Beate Vieth
}
