% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{evaluateROC}
\alias{evaluateROC}
\title{Receiver operator characteristics of simulation results}
\usage{
evaluateROC(simRes,
alpha.type=c("adjusted","raw"),
MTC=c('BY', 'BH', 'Storey', 'IHW',
'holm', 'hochberg', 'hommel', 'bonferroni'),
alpha.nominal = 0.1,
target.by=c("lfc", "effectsize"),
delta=0)
}
\arguments{
\item{simRes}{The result from \code{\link{simulateDE}}.}

\item{alpha.type}{A string to represent the way to call DE genes.
Available options are \code{"adjusted"} i.e. applying multiple testing correction and
\code{"raw"} i.e. using p-values. Default is \code{"adjusted"}.}

\item{MTC}{Multiple testing correction method to use. Available options are
1) see \link[stats]{p.adjust.methods},
2) Storey's qvalue see \link[qvalue]{qvalue} and
3) Independent Hypothesis Weighting considering mean expression as covariate (see \link[IHW]{ihw}).
Default is \code{BY}, i.e. Benjamini-Yekutieli FDR correction method.}

\item{alpha.nominal}{The nomial value of significance. Default is 0.1.}

\item{target.by}{A string to specify the method to define "biologically important" DE genes.
Available options are (1) \code{"lfc"}: interesting genes are defined by absolute log2 fold changes.
(2) \code{"effectsize"}: interesting genes are defined by
absolute log2 fold changes divided by the square root of 1/(mean+dispersion).}

\item{delta}{A threshold used for defining "biologically important" genes.
Genes with absolute log2 fold changes (when target.by is "lfc")
or effect sizes (when target.by is "effectsize") greater than this value
are deemed DE in error rates calculations.
If \code{delta=0} then no threshold is applied. See "Details" for more description.}
}
\value{
A list with the following entries:
\item{Performance}{The output of \code{\link[iCOBRA]{calculate_performance}} of aspect "fdrtprcurve" calculating the proportions of TN, FN, TP and FP and related rates which uses \code{\link[ROCR]{prediction}} and \code{\link[ROCR]{performance}} internally.}
\item{FDR_TPR_Thres}{The output of \code{\link[iCOBRA]{calculate_performance}} of aspect "fdrtpr" calculating the proportions of TN, FN, TP and FP and associated TPR and FDR for each nominal level from 0.01 to 1 in steps of 0.01.}
\item{TPRvsPPV_AUC}{The area under the curve for TPR versus PPV. This is also split up by group comparisons into lists with length equal to number of simulations.}
\item{TPRvsFPR_AUC}{The area under the curve for TPR versus FPR, i.e. classical ROC-curve. This is also split up by group comparisons into lists with length equal to number of simulations.}
\item{TPRvsFDR_pAUC_conv,TPRvsFDR_pAUC_lib}{The partial area under the curve for TPR versus FDR, once for conservative and once for liberal nominal FDR level (see Details section). This is also split up by group comparisons into lists with length equal to number of simulations.}
\item{MCC_conv,MCC_lib}{The Matthews Correlation Coefficient, once for conservative and once for liberal nominal FDR level (see Details section). This is also split up by group comparisons into lists with length equal to number of simulations.}
\item{F1score_conv,F1score_lib}{The precision-recall F measure, once for conservative and once for liberal nominal FDR level (see Details section). This is also split up by group comparisons into lists with length equal to number of simulations.}
\item{alpha.type - .. - delta}{Reiterating chosen evaluation parameters.}
}
\description{
This function takes the simulation output from \code{\link{simulateDE}}
and computes receiver operator characteristics (ROC) and area under the curve values (AUC).
}
\examples{
\dontrun{
data(kolodziejczk_simDE)
eval.roc <- evaluateROC(simRes = kolodziejczk_simDE)
}
}
\seealso{
\code{\link{estimateParam}} for parameter estimation needed for simulations,
\code{\link{Setup}} for setting up simulation parameters and
\code{\link{simulateDE}} for simulating differential expression.
}
\author{
Beate Vieth
}
