% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{evaluateSim}
\alias{evaluateSim}
\title{Compute the performance related metrics from simulation results}
\usage{
evaluateSim(simRes)
}
\arguments{
\item{simRes}{The result from \code{\link{simulateDE}} or \code{\link{simulateFlow}}.}
}
\value{
A list with the following entries:
\item{Log2FoldChange}{The root mean square error and absolute mean error of
log2 fold change differences between estimated LFC and simulated LFC. Furthermore, the fraction of missing estimated LFC.}
\item{SizeFactors}{The median absolute deviation between the estimated and true size factors,
the residual error of a robust linear model and the ratio between estimated and true size factors of the two groups.}
\item{Clustering}{The adjusted rand index between the true group assignment and derived group assignment by clustering (only for \code{\link{simulateFlow}}).}
}
\description{
This function takes the simulation output from \code{\link{simulateDE}}
and computes several metrics that give an indication of the simulation setup performance.
}
\examples{
\dontrun{
## not yet
}
}
\seealso{
\code{\link{estimateParam}} for negative binomial parameters,
\code{\link{SimSetup}} and
\code{\link{DESetup}} for setting up simulation parameters and
\code{\link{simulateDE}} and \code{\link{simulateFlow}} for simulating differential expression.
}
\author{
Beate Vieth
}
