% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{evaluateSim}
\alias{evaluateSim}
\title{Compute the performance related metrics from simulation results.}
\usage{
evaluateSim(simRes, timing=TRUE)
}
\arguments{
\item{simRes}{The result from \code{\link{simulateDE}} or \code{\link{simulateFlow}}.}

\item{timing}{A logical vector indicating whether to summarise computational time of simulation run.
Default is \code{TRUE}.}
}
\value{
A list with the following entries:
\item{Log2FoldChange}{The absolute mean error (\code{MAE}), root mean square error (\code{RMSE})
and root mean square residual error of a robust linear model (\code{rRMSE}, \code{\link[MASS]{rlm}})
of log2 fold change differences between estimated LFC and simulated LFC.
Furthermore, the fraction of missing entries (\code{NAFraction}) for MAE annd RMSE.}
\item{SizeFactors}{The median absolute deviation (MAD) between the estimated and true size factors,
the root mean square residual error of a robust linear model (rRMSE, \code{\link[MASS]{rlm}}) and
the ratio between estimated and true size factors of the two groups (\code{GroupX}).}
\item{Clustering}{The adjusted rand index between the true group assignment and derived group assignment
by clustering (only for \code{\link{simulateFlow}}).}
}
\description{
This function takes the simulation output from \code{\link{simulateDE}} or \code{\link{simulateFlow}}
and computes several metrics that give an indication of the simulation setup performance.
}
\examples{
\dontrun{
## using example data set
data(kolodziejczk_simDE)
eval.sim <- evaluateSim(simRes = kolodziejczk_simDE, timing = T)
}
}
\seealso{
\code{\link{estimateParam}} for negative binomial parameters,
\code{\link{SimSetup}} and
\code{\link{DESetup}} for setting up simulation parameters and
\code{\link{simulateDE}} for simulating differential expression and
\code{\link{simulateFlow}} for simulating workflow.
}
\author{
Beate Vieth
}
