% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotEvalDE}
\alias{plotEvalDE}
\title{Visualize statistical power assessment}
\usage{
plotEvalDE(evalRes, rate=c('marginal', 'stratified'),
                   quick=TRUE, annot=TRUE)
}
\arguments{
\item{evalRes}{The output of \code{\link{evaluateDE}}.}

\item{rate}{Character vector defining whether the marginal or condtional rates should be plotted. Conditional depends on the choice of stratify.by in \code{\link{evaluateDE}}.}

\item{quick}{A logical vector. If \code{TRUE}, the TPR and FDR are only plotted. If \code{FALSE}, then all rates are plotted.}

\item{annot}{A logical vector. If \code{TRUE}, a short figure legend under the plot is included.}
}
\value{
A ggplot object.
}
\description{
This function plots the results of \code{\link{evaluateDE}} for assessing the error rates and sample size requirements.
}
\examples{
\dontrun{
## using example data set
eval.de <- evaluateDE(simRes = kolodziejczk_simDE)
plotEvalDE(evalRes=eval.de, rate ="marginal", quick=T, annot=T)
plotEvalDE(evalRes=eval.de, rate ="stratified", quick=T, annot=T)
}
}
\author{
Beate Vieth
}
