% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotEvalDist}
\alias{plotEvalDist}
\title{Visualize distribution assessment}
\usage{
plotEvalDist(evalDist, annot=TRUE)
}
\arguments{
\item{evalDist}{The output of \code{\link{evaluateDist}}.}

\item{annot}{A logical vector. If \code{TRUE}, a short description of the plot is included.}
}
\value{
A ggplot object.
}
\description{
This function plots the results of \code{\link{evaluateDist}} to assess goodness-of-fit testing.
}
\examples{
\dontrun{
## using example data set
data(kolodziejczk_cnts)
evaldist <- evaluateDist(countData = kolodziejczk_cnts,
RNAseq = "singlecell", normalisation="scran",
frac.genes=1, min.meancount = 0.1,
max.dropout=0.7, min.libsize=1000,
verbose = TRUE)
plotEvalDist(evaldist, annot = TRUE)
}
}
\author{
Beate Vieth, Ines Hellmann
}
