% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plotTime}
\alias{plotTime}
\title{Visualize computational time of simulations}
\usage{
plotTime(simRes, Table=TRUE, annot=TRUE)
}
\arguments{
\item{simRes}{The output of \code{\link{simulateDE}}.}

\item{Table}{A logical vector. If \code{TRUE}, a table of average computational running time per step and sample size is printed additionally.}

\item{annot}{A logical vector. If \code{TRUE}, a short figure legend under the plot is included.}
}
\value{
A ggplot object.
}
\description{
This function plots the computational running time of simulations.
}
\examples{
\dontrun{
## using example data set
plotTime(simRes = kolodziejczk_simDE)
}
}
\author{
Beate Vieth
}
