% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{printEvalDE}
\alias{printEvalDE}
\title{Summary table of power assessment}
\usage{
printEvalDE(evalRes)
}
\arguments{
\item{evalRes}{The result object from \code{\link{evaluateDE}}.}
}
\value{
A matrix of results per sample size considered (rows). Columns include sample size, specified nomial significance level (for FDR or p-values),  the actual marginal error rate and marginal TPR.
}
\description{
This function takes as input a result object from \code{\link{evaluateDE}}
and prints out a table to summarize important error-rates-related quantities.
The results are marginalized, meaning that they are averaged quantities
over all strata and simulations.
This provides a quick view of the statistical power analysis per sample size setup.
}
\examples{
\dontrun{
## for example evaluation result see \code{\link{evaluateDE}}
}
}
\seealso{
\code{\link{simulateDE}}, \code{\link{evaluateDE}}
}
\author{
Beate Vieth
}
