% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim.R
\name{simulateDE}
\alias{simulateDE}
\title{Simulate Differential Expression}
\usage{
simulateDE(n1=c(20,50,100), n2=c(30,60,120),
sim.settings,
DEmethod,
normalisation,
Preclust=FALSE,
Preprocess = NULL,
spikeIns=FALSE,
NCores=NULL,
verbose=TRUE)
}
\arguments{
\item{n1, n2}{Integer vectors specifying the number of biological replicates in each group. Default values are n1=c(20,50,100) and n2=c(30,60,120).}

\item{sim.settings}{This object specifies the simulation setup. This must be the return object from \code{\link{SimSetup}}.}

\item{DEmethod}{A character vector specifying the DE detection method to be used.
Available options are: limma-trend, limma-voom, edgeR-LRT, edgeR-QL, DESeq2,
ROTS, baySeq, NOISeq, EBSeq, DSS, MAST, scde, BPSC, scDD, monocle.}

\item{normalisation}{Normalisation method to use.
Available options are:}

\item{Preclust}{Whether to run a  hierarchical clustering prior to normalisation. Default is \code{FALSE}. This is implemented for scran and SCnorm.
For details, see \code{\link[scran]{quickCluster}}.}

\item{Preprocess}{A character vector specifying the gene filtering method to be used
prior to normalisation. Default is \code{NULL}, i.e. no filtering.
Availabe options are: scImpute, DrImpute, CountFilter, FreqFilter.}

\item{spikeIns}{Logical value to indicate whether to simulate spike-ins. Default is \code{FALSE}.}

\item{NCores}{integer positive number of cores for parallel processing, default is \code{NULL}, ie 1 core.}

\item{verbose}{Logical value to indicate whether to show progress report of simulations. Default is \code{TRUE}.}
}
\value{
A list with the following fields.
\item{pvalue, fdr}{3D array (ngenes * N * nsims) for p-values and FDR from each simulation.
Note that FDR values will be empty and the calculation will be done by \code{\link{evaluateDE}} whenever applicable.}
\item{mu,disp,dropout}{3D (ngenes * N * nsims) array for mean, dispersion and dropout of library size factor normalized read counts.}
\item{elfc,rlfc}{3D array (ngenes * N * nsims) for log2 fold changes (LFC):
elfc is for the DE tool estimated LFC; rlfc is for the LFC estimated from the normalised read counts.}
\item{sf.values,gsf.values}{3D array (ngenes * N * nsims) for size factor estimates.
Global estimates per sample in sf.values; Gene- and sample-wise estimates in gsf.values only for SCnorm normalisation.}
\item{sim.settings}{The input sim.settings to which the specifications of \code{simulateDE} is added.}
\item{time.taken}{The time taken for each simulation, given for preprocessing, normalisation, clustering, differential expression testing and moment estimation.}
}
\description{
simulateDE is the main function to simulate differential expression for RNA-seq experiments.
The simulation parameters are specified with \code{\link{SimSetup}}.
The user needs to specify furthermore
the number of samples per group, preprocessing, normalisation and differential testing method.
There is also the option to consider spike-ins. \cr
The return object contains DE test results from all simulations as well as descriptive statistics.
The error matrix calculations will be conducted with \code{\link{evaluateDE}}.\cr
}
\examples{
\dontrun{
# download count table
githubURL <- "https://github.com/bvieth/powsimRData/raw/master/data-raw/kolodziejczk_cnts.rda"
download.file(url = githubURL, destfile= "kolodziejczk_cnts.rda", method = "wget")
load('kolodziejczk_cnts.rda')
kolodziejczk_cnts <- kolodziejczk_cnts[, grep('standard', colnames(kolodziejczk_cnts))]
## estimate NB parameters:
TwoiLIF.params =
## define DE settings:
desettings <- DESetup(ngenes=10000,
nsims=25, p.DE=0.2,
LFC=function(x) sample(c(-1,1), size=x,replace=TRUE)*rgamma(x, 3, 3))
## define simulation settings for Kolodziejczk:
simsettings <- SimSetup(desetup=desettings, params=TwoiLIF.params, size.factors='given')
## run simulations:
## if parallel computation unavailable, consider ROTS as DEmethod
}
}
\seealso{
\code{\link{estimateParam}},  \code{\link{insilicoNBParam}} for negative binomial parameter specifications;\cr
 \code{\link{DESetup}}, \code{\link{SimSetup}} for simulation setup
}
\author{
Beate Vieth
}
